requirements.txt
requests~=2.26.0
python-telegram-bot
backoff~=1.11.1
fastapi~=0.70.0
